books/pyproject.toml
[project]
name = "books"
version = "0.1.0"
description = "Books microservice for bookstore"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "cloudinary>=1.36.0",
    "python-multipart>=0.0.6"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "pytest-mock>=3.10.0"
]
